+++ example/lib/main.dart
@@ -116,6 +116,18 @@ class _ExampleAppState extends State<ExampleApp> {
                 onDateTimeChanged: _onDateTimeChanged,
               ),
             ),
+            SizedBox(
+              width: 350,
+              child: CupertinoCalendar(
+                minimumDateTime: _minimumDateTime,
+                maximumDateTime: _maximumDateTime,
+                initialDateTime: _selectedDateTime,
+                selectableDayPredicate: (date) => date.day % 2 == 0,
+                timeLabel: 'Ends',
+                mode: CupertinoCalendarMode.dateTimeWeek,
+                onDateTimeChanged: _onDateTimeChanged,
+              ),
+            ),
             const Spacer(),
             Row(
               mainAxisAlignment: MainAxisAlignment.center,