+++ lib/src/calendar/picker/cupertino_calendar_picker.dart
@@ -8,7 +8,7 @@ import 'package:flutter/material.dart';
 
 class CupertinoCalendarPicker extends StatefulWidget {
   const CupertinoCalendarPicker({
-    required this.initialMonth,
+    required this.initialDate,
     required this.currentDateTime,
     required this.minimumDateTime,
     required this.maximumDateTime,
@@ -18,9 +18,11 @@ class CupertinoCalendarPicker extends StatefulWidget {
     required this.onDateChanged,
     required this.onTimeChanged,
     required this.onDisplayedMonthChanged,
+    required this.onDisplayedWeekChanged,
     required this.onYearPickerChanged,
     required this.weekdayDecoration,
     required this.monthPickerDecoration,
+    required this.weekPickerDecoration,
     required this.headerDecoration,
     required this.mainColor,
     required this.mode,
@@ -33,7 +35,7 @@ class CupertinoCalendarPicker extends StatefulWidget {
     super.key,
   });
 
-  final DateTime initialMonth;
+  final DateTime initialDate;
   final DateTime currentDateTime;
   final DateTime minimumDateTime;
   final DateTime maximumDateTime;
@@ -42,9 +44,11 @@ class CupertinoCalendarPicker extends StatefulWidget {
   final ValueChanged<DateTime> onDateChanged;
   final ValueChanged<DateTime> onTimeChanged;
   final ValueChanged<DateTime> onDisplayedMonthChanged;
+  final ValueChanged<DateTime> onDisplayedWeekChanged;
   final ValueChanged<DateTime> onYearPickerChanged;
   final CalendarWeekdayDecoration weekdayDecoration;
   final CalendarMonthPickerDecoration monthPickerDecoration;
+  final CalendarWeekPickerDecoration weekPickerDecoration;
   final CalendarHeaderDecoration headerDecoration;
   final CalendarFooterDecoration footerDecoration;
   final Color mainColor;
@@ -61,9 +65,9 @@ class CupertinoCalendarPicker extends StatefulWidget {
 }
 
 class CupertinoCalendarPickerState extends State<CupertinoCalendarPicker> {
-  late DateTime _currentMonth;
+  late DateTime _currentDate;
   late DateTime _selectedDateTime;
-  late PageController _monthPageController;
+  late PageController _pageController;
   late CupertinoCalendarViewMode _previousViewMode;
   late CupertinoCalendarViewMode _viewMode;
   late GlobalKey<CustomCupertinoDatePickerDateTimeState> _timePickerKey;
@@ -78,27 +82,42 @@ class CupertinoCalendarPickerState extends State<CupertinoCalendarPicker> {
   @override
   void initState() {
     super.initState();
-    _currentMonth = widget.initialMonth;
-    final int monthDelta = DateUtils.monthDelta(
-      widget.minimumDateTime,
-      _currentMonth,
-    );
-    _monthPageController = PageController(initialPage: monthDelta);
+    _currentDate = widget.initialDate;
     _previousViewMode = CupertinoCalendarViewMode.monthPicker;
     _viewMode = CupertinoCalendarViewMode.monthPicker;
-    _selectedDateTime = widget.selectedDateTime;
     _timePickerKey = GlobalKey();
+    _selectedDateTime = widget.selectedDateTime;
+    if (widget.mode == CupertinoCalendarMode.date ||
+        widget.mode == CupertinoCalendarMode.dateTime) {
+      _initialisePageController();
+    }
+  }
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    if (widget.mode == CupertinoCalendarMode.dateWeek ||
+        widget.mode == CupertinoCalendarMode.dateTimeWeek) {
+      _initialisePageController();
+    }
   }
 
   @override
   void didUpdateWidget(CupertinoCalendarPicker oldWidget) {
     super.didUpdateWidget(oldWidget);
-    final DateTime initialMonth = widget.initialMonth;
-    final DateTime oldInitialMonth = oldWidget.initialMonth;
-    if (initialMonth != oldInitialMonth && initialMonth != _currentMonth) {
+    final DateTime initialDate = widget.initialDate;
+    final DateTime oldInitialDate = oldWidget.initialDate;
+    if (initialDate != oldInitialDate && initialDate != _currentDate) {
       // We can't interrupt this widget build with a scroll, so do it next frame
       WidgetsBinding.instance.addPostFrameCallback(
-        (Duration timeStamp) => _showMonth(widget.initialMonth, jump: true),
+        (Duration timeStamp) => switch (widget.mode) {
+          CupertinoCalendarMode.date ||
+          CupertinoCalendarMode.dateTime =>
+            _showMonth(widget.initialDate, jump: true),
+          CupertinoCalendarMode.dateWeek ||
+          CupertinoCalendarMode.dateTimeWeek =>
+            _showWeek(widget.initialDate, jump: true),
+        },
       );
     }
 
@@ -110,7 +129,7 @@ class CupertinoCalendarPickerState extends State<CupertinoCalendarPicker> {
   /// Earliest allowable month.
   bool get _isDisplayingFirstMonth {
     final DateTime minimumDate = widget.minimumDateTime;
-    return !_currentMonth.isAfter(
+    return !_currentDate.isAfter(
       DateTime(minimumDate.year, minimumDate.month),
     );
   }
@@ -118,9 +137,50 @@ class CupertinoCalendarPickerState extends State<CupertinoCalendarPicker> {
   /// Latest allowable month.
   bool get _isDisplayingLastMonth {
     final DateTime maximumDate = widget.maximumDateTime;
-    return !_currentMonth.isBefore(
+    return !_currentDate.isBefore(
       DateTime(maximumDate.year, maximumDate.month),
     );
+  }
+
+  /// Earliest allowable week.
+  bool get _isDisplayingFirstWeek {
+    final DateTime minimumDate = widget.minimumDateTime;
+    return _currentDate.isSameWeekAs(
+      minimumDate,
+      widget.firstDayOfWeekIndex ??
+          context.materialLocalization.firstDayOfWeekIndex,
+    );
+  }
+
+  /// Latest allowable week.
+  bool get _isDisplayingLastWeek {
+    final DateTime maximumDate = widget.maximumDateTime;
+    return _currentDate.isSameWeekAs(
+      maximumDate,
+      widget.firstDayOfWeekIndex ??
+          context.materialLocalization.firstDayOfWeekIndex,
+    );
+  }
+
+  void _initialisePageController() {
+    final int delta = switch (widget.mode) {
+      CupertinoCalendarMode.date ||
+      CupertinoCalendarMode.dateTime =>
+        DateUtils.monthDelta(
+          widget.minimumDateTime,
+          _currentDate,
+        ),
+      CupertinoCalendarMode.dateWeek ||
+      CupertinoCalendarMode.dateTimeWeek =>
+        PackageDateUtils.weekDelta(
+          widget.minimumDateTime,
+          _currentDate,
+          widget.firstDayOfWeekIndex ??
+              context.materialLocalization.firstDayOfWeekIndex,
+        ),
+    };
+
+    _pageController = PageController(initialPage: delta);
   }
 
   void _handleMonthPageChanged(int monthPage) {
@@ -129,17 +189,34 @@ class CupertinoCalendarPickerState extends State<CupertinoCalendarPicker> {
       final DateTime monthDate =
           DateUtils.addMonthsToMonthDate(minimumDate, monthPage);
       final bool isCurrentMonth =
-          DateUtils.isSameMonth(_currentMonth, monthDate);
+          DateUtils.isSameMonth(_currentDate, monthDate);
       if (!isCurrentMonth) {
-        _currentMonth = DateTime(monthDate.year, monthDate.month);
-        widget.onDisplayedMonthChanged(_currentMonth);
+        _currentDate = DateTime(monthDate.year, monthDate.month);
+        widget.onDisplayedMonthChanged(_currentDate);
+      }
+    });
+  }
+
+  void _handleWeekPageChanged(int weekPage) {
+    setState(() {
+      final DateTime minimumDate = widget.minimumDateTime;
+      final DateTime weekDate =
+          DateUtils.addDaysToDate(minimumDate, weekPage * 7);
+      final bool isCurrentWeek = _currentDate.isSameWeekAs(
+        weekDate,
+        widget.firstDayOfWeekIndex ??
+            context.materialLocalization.firstDayOfWeekIndex,
+      );
+      if (!isCurrentWeek) {
+        _currentDate = DateTime(weekDate.year, weekDate.month, weekDate.day);
+        widget.onDisplayedWeekChanged(_currentDate);
       }
     });
   }
 
   void _handleNextMonth() {
     if (!_isDisplayingLastMonth) {
-      _monthPageController.nextPage(
+      _pageController.nextPage(
         duration: monthScrollDuration,
         curve: Curves.ease,
       );
@@ -148,20 +225,75 @@ class CupertinoCalendarPickerState extends State<CupertinoCalendarPicker> {
 
   void _handlePreviousMonth() {
     if (!_isDisplayingFirstMonth) {
-      _monthPageController.previousPage(
+      _pageController.previousPage(
         duration: monthScrollDuration,
         curve: Curves.ease,
       );
     }
   }
 
+  void _handleNextWeek() {
+    if (!_isDisplayingLastWeek) {
+      _pageController.nextPage(
+        duration: monthScrollDuration,
+        curve: Curves.ease,
+      );
+    }
+  }
+
+  void _handlePreviousWeek() {
+    if (!_isDisplayingFirstWeek) {
+      _pageController.previousPage(
+        duration: monthScrollDuration,
+        curve: Curves.ease,
+      );
+    }
+  }
+
+  VoidCallback? get _nextChevronCallback => switch (widget.mode) {
+        CupertinoCalendarMode.date ||
+        CupertinoCalendarMode.dateTime =>
+          _isDisplayingLastMonth ? null : _handleNextMonth,
+        CupertinoCalendarMode.dateWeek ||
+        CupertinoCalendarMode.dateTimeWeek =>
+          _isDisplayingLastWeek ? null : _handleNextWeek
+      };
+
+  VoidCallback? get _previousChevronCallback => switch (widget.mode) {
+        CupertinoCalendarMode.date ||
+        CupertinoCalendarMode.dateTime =>
+          _isDisplayingFirstMonth ? null : _handlePreviousMonth,
+        CupertinoCalendarMode.dateWeek ||
+        CupertinoCalendarMode.dateTimeWeek =>
+          _isDisplayingFirstWeek ? null : _handlePreviousWeek
+      };
+
   void _showMonth(DateTime month, {bool jump = false}) {
     final int monthPage = DateUtils.monthDelta(widget.minimumDateTime, month);
     if (jump) {
-      _monthPageController.jumpToPage(monthPage);
+      _pageController.jumpToPage(monthPage);
     } else {
-      _monthPageController.animateToPage(
+      _pageController.animateToPage(
         monthPage,
+        duration: monthScrollDuration,
+        curve: Curves.ease,
+      );
+    }
+  }
+
+  void _showWeek(DateTime week, {bool jump = false}) {
+    final int weekPage = PackageDateUtils.weekDelta(
+      widget.minimumDateTime,
+      week,
+      widget.firstDayOfWeekIndex ??
+          context.materialLocalization.firstDayOfWeekIndex,
+    );
+
+    if (jump) {
+      _pageController.jumpToPage(weekPage);
+    } else {
+      _pageController.animateToPage(
+        weekPage,
         duration: monthScrollDuration,
         curve: Curves.ease,
       );
@@ -192,7 +324,7 @@ class CupertinoCalendarPickerState extends State<CupertinoCalendarPicker> {
     widget.onYearPickerChanged(date);
   }
 
-  void _onMonthDateChanged(DateTime dateTime) {
+  void _onDateChanged(DateTime dateTime) {
     _selectedDateTime = _selectedDateTime.copyWith(
       year: dateTime.year,
       month: dateTime.month,
@@ -250,11 +382,10 @@ class CupertinoCalendarPickerState extends State<CupertinoCalendarPicker> {
           const SizedBox(height: 13.0),
           CupertinoPickerAnimatedCrossFade(
             firstChild: CalendarHeader(
-              currentMonth: _currentMonth,
-              onNextMonthIconTapped:
-                  _isDisplayingLastMonth ? null : _handleNextMonth,
-              onPreviousMonthIconTapped:
-                  _isDisplayingFirstMonth ? null : _handlePreviousMonth,
+              mode: widget.mode,
+              currentMonth: _currentDate,
+              onNextSegmentIconTapped: _nextChevronCallback,
+              onPreviousSegmentIconTapped: _previousChevronCallback,
               onYearPickerStateChanged: _toggleYearPicker,
               decoration: widget.headerDecoration,
             ),
@@ -276,20 +407,38 @@ class CupertinoCalendarPickerState extends State<CupertinoCalendarPicker> {
                     decoration: widget.weekdayDecoration,
                     firstDayOfWeekIndex: widget.firstDayOfWeekIndex,
                   ),
-                  CalendarMonthPicker(
-                    monthPageController: _monthPageController,
-                    onMonthPageChanged: _handleMonthPageChanged,
-                    currentDate: widget.currentDateTime,
-                    displayedMonth: _currentMonth,
-                    minimumDate: widget.minimumDateTime,
-                    maximumDate: widget.maximumDateTime,
-                    selectableDayPredicate: widget.selectableDayPredicate,
-                    selectedDate: widget.selectedDateTime,
-                    onChanged: _onMonthDateChanged,
-                    decoration: widget.monthPickerDecoration,
-                    mainColor: widget.mainColor,
-                    firstDayOfWeekIndex: widget.firstDayOfWeekIndex,
-                  ),
+                  switch (widget.mode) {
+                    CupertinoCalendarMode.date ||
+                    CupertinoCalendarMode.dateTime =>
+                      CalendarMonthPicker(
+                        monthPageController: _pageController,
+                        onMonthPageChanged: _handleMonthPageChanged,
+                        currentDate: widget.currentDateTime,
+                        minimumDate: widget.minimumDateTime,
+                        maximumDate: widget.maximumDateTime,
+                        selectableDayPredicate: widget.selectableDayPredicate,
+                        selectedDate: widget.selectedDateTime,
+                        onChanged: _onDateChanged,
+                        decoration: widget.monthPickerDecoration,
+                        mainColor: widget.mainColor,
+                        firstDayOfWeekIndex: widget.firstDayOfWeekIndex,
+                      ),
+                    CupertinoCalendarMode.dateWeek ||
+                    CupertinoCalendarMode.dateTimeWeek =>
+                      CalendarWeekPicker(
+                        weekPageController: _pageController,
+                        onWeekPageChanged: _handleWeekPageChanged,
+                        currentDate: widget.currentDateTime,
+                        minimumDate: widget.minimumDateTime,
+                        maximumDate: widget.maximumDateTime,
+                        selectableDayPredicate: widget.selectableDayPredicate,
+                        selectedDate: widget.selectedDateTime,
+                        onChanged: _onDateChanged,
+                        decoration: widget.weekPickerDecoration,
+                        mainColor: widget.mainColor,
+                        firstDayOfWeekIndex: widget.firstDayOfWeekIndex,
+                      )
+                  },
                 ],
               ),
               secondChild: Padding(
@@ -312,7 +461,7 @@ class CupertinoCalendarPickerState extends State<CupertinoCalendarPicker> {
                       ),
                       mode: CupertinoDatePickerMode.monthYear,
                       onDateTimeChanged: _onYearPickerChanged,
-                      initialDateTime: _currentMonth,
+                      initialDateTime: _currentDate,
                     ),
                   CupertinoCalendarViewMode.timePicker =>
                     CupertinoTimePickerWheel(
@@ -331,7 +480,8 @@ class CupertinoCalendarPickerState extends State<CupertinoCalendarPicker> {
               ),
             ),
           ),
-          if (widget.mode == CupertinoCalendarMode.dateTime)
+          if (widget.mode == CupertinoCalendarMode.dateTime ||
+              widget.mode == CupertinoCalendarMode.dateTimeWeek)
             CupertinoPickerAnimatedCrossFade(
               firstChild: CalendarFooter(
                 decoration: widget.footerDecoration,
@@ -362,7 +512,7 @@ class CupertinoCalendarPickerState extends State<CupertinoCalendarPicker> {
 
   @override
   void dispose() {
-    _monthPageController.dispose();
+    _pageController.dispose();
     super.dispose();
   }
 }