+++ lib/src/calendar/week_picker/calendar_week_picker_decoration.dart
@@ -5,18 +5,18 @@
 import 'package:cupertino_calendar_picker/src/src.dart';
 import 'package:flutter/cupertino.dart';
 
-/// A decoration class for the calendar's month picker.
-class CalendarMonthPickerDecoration {
-  /// Creates a calendar's month picker decoration class with default values
+/// A decoration class for the calendar's week picker.
+class CalendarWeekPickerDecoration {
+  /// Creates a calendar's week picker decoration class with default values
   /// for non-provided parameters.
-  factory CalendarMonthPickerDecoration({
+  factory CalendarWeekPickerDecoration({
     CalendarPickerDefaultDayStyle? defaultDayStyle,
     CalendarPickerCurrentDayStyle? currentDayStyle,
     CalendarPickerSelectedDayStyle? selectedDayStyle,
     CalendarPickerSelectedCurrentDayStyle? selectedCurrentDayStyle,
     CalendarPickerDisabledDayStyle? disabledDayStyle,
   }) {
-    return CalendarMonthPickerDecoration._(
+    return CalendarWeekPickerDecoration._(
       defaultDayStyle: defaultDayStyle,
       currentDayStyle: currentDayStyle,
       selectedDayStyle: selectedDayStyle,
@@ -25,7 +25,7 @@ class CalendarMonthPickerDecoration {
     );
   }
 
-  const CalendarMonthPickerDecoration._({
+  const CalendarWeekPickerDecoration._({
     this.defaultDayStyle,
     this.currentDayStyle,
     this.selectedDayStyle,
@@ -33,22 +33,22 @@ class CalendarMonthPickerDecoration {
     this.disabledDayStyle,
   });
 
-  /// Creates a calendar's month picker decoration class with default values
+  /// Creates a calendar's week picker decoration class with default values
   /// for non-provided parameters.
   ///
   /// Applies the [CupertinoDynamicColor.resolve] method for colors.
   ///
   /// [mainColor] is used only if any other color is not provided.
-  factory CalendarMonthPickerDecoration.withDynamicColor(
-    BuildContext context, {
-    Color? mainColor,
-    CalendarPickerDefaultDayStyle? defaultDayStyle,
-    CalendarPickerCurrentDayStyle? currentDayStyle,
-    CalendarPickerSelectedDayStyle? selectedDayStyle,
-    CalendarPickerSelectedCurrentDayStyle? selectedCurrentDayStyle,
-    CalendarPickerDisabledDayStyle? disabledDayStyle,
-  }) {
-    return CalendarMonthPickerDecoration(
+  factory CalendarWeekPickerDecoration.withDynamicColor(
+      BuildContext context, {
+        Color? mainColor,
+        CalendarPickerDefaultDayStyle? defaultDayStyle,
+        CalendarPickerCurrentDayStyle? currentDayStyle,
+        CalendarPickerSelectedDayStyle? selectedDayStyle,
+        CalendarPickerSelectedCurrentDayStyle? selectedCurrentDayStyle,
+        CalendarPickerDisabledDayStyle? disabledDayStyle,
+      }) {
+    return CalendarWeekPickerDecoration(
       defaultDayStyle: defaultDayStyle ??
           CalendarPickerDefaultDayStyle.withDynamicColor(context),
       currentDayStyle: currentDayStyle ??
@@ -74,39 +74,39 @@ class CalendarMonthPickerDecoration {
   }
 
   /// The [CalendarPickerDefaultDayStyle] of the
-  /// calendar's month picker default day.
+  /// calendar's week picker default day.
   final CalendarPickerDefaultDayStyle? defaultDayStyle;
 
   /// The [CalendarPickerCurrentDayStyle] of the
-  /// calendar's month picker current day.
+  /// calendar's week picker current day.
   final CalendarPickerCurrentDayStyle? currentDayStyle;
 
   /// The [CalendarPickerSelectedDayStyle] of the
-  /// calendar's month picker selected day.
+  /// calendar's week picker selected day.
   final CalendarPickerSelectedDayStyle? selectedDayStyle;
 
   /// The [CalendarPickerSelectedCurrentDayStyle] of the
-  /// calendar's month picker selected current day.
+  /// calendar's week picker selected current day.
   final CalendarPickerSelectedCurrentDayStyle? selectedCurrentDayStyle;
 
   /// The [CalendarPickerDisabledDayStyle] of the
-  /// calendar's month picker disabled day.
+  /// calendar's week picker disabled day.
   final CalendarPickerDisabledDayStyle? disabledDayStyle;
 
   /// Creates a copy of the class with the provided parameters.
-  CalendarMonthPickerDecoration copyWith({
+  CalendarWeekPickerDecoration copyWith({
     CalendarPickerDefaultDayStyle? defaultDayStyle,
     CalendarPickerCurrentDayStyle? currentDayStyle,
     CalendarPickerSelectedDayStyle? selectedDayStyle,
     CalendarPickerSelectedCurrentDayStyle? selectedCurrentDayStyle,
     CalendarPickerDisabledDayStyle? disabledDayStyle,
   }) {
-    return CalendarMonthPickerDecoration(
+    return CalendarWeekPickerDecoration(
       defaultDayStyle: defaultDayStyle ?? defaultDayStyle,
       currentDayStyle: currentDayStyle ?? currentDayStyle,
       selectedDayStyle: selectedDayStyle ?? this.selectedDayStyle,
       selectedCurrentDayStyle:
-          selectedCurrentDayStyle ?? selectedCurrentDayStyle,
+      selectedCurrentDayStyle ?? selectedCurrentDayStyle,
       disabledDayStyle: disabledDayStyle ?? this.disabledDayStyle,
     );
   }